# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Система заявок на производстве",
        "contact": {
            "name": "Корецкий К.В.",
            "url": "https://github.com/korzck",
            "email": "konstantin.koretskiy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/items": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get list of all items",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "filter by min price",
                        "name": "min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "text",
                        "description": "filter by max price",
                        "name": "max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "text",
                        "description": "filter by material (wood/metal)",
                        "name": "material",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ItemsSwagger"
                        }
                    }
                }
            }
        },
        "/items/image": {
            "post": {
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Upload s3 file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "upload file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "metadata",
                        "name": "metadata",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ImageSwagger"
                        }
                    }
                }
            }
        },
        "/items/post": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create item",
                "parameters": [
                    {
                        "description": "Item object",
                        "name": "itemPrototype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ItemPrototype"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ItemPrototype"
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "item id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ItemModel"
                        }
                    }
                }
            }
        },
        "/items/{id}/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete item by id",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "text",
                        "description": "item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/items/{id}/post": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Post item to current order",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "text",
                        "description": "item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.Order"
                        }
                    }
                }
            }
        },
        "/items/{id}/put": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Change item",
                "parameters": [
                    {
                        "description": "Item object",
                        "name": "itemPrototype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.ItemPrototype"
                        }
                    },
                    {
                        "type": "integer",
                        "format": "text",
                        "description": "item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "userCreds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.UserCreds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.UserSwagger"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "type": "string",
                        "default": "token=xxx",
                        "description": "token",
                        "name": "Cookie",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get list of all orders",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "min date",
                        "name": "min_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "text",
                        "description": "max date",
                        "name": "max_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "text",
                        "description": "order status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web_internal_models.Order"
                            }
                        }
                    }
                }
            }
        },
...
        "/orders/items/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete item from current order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.OrderSwagger"
                        }
                    }
                }
            }
        },
...
        "/orders/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "text",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.OrderSwagger"
                        }
                    }
                }
            }
        },
        "/orders/{id}/approve": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Approve or decline order",
                "parameters": [
                    {
                        "description": "Order status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.OrderStatusSwagger"
                        }
                    },
                    {
                        "type": "string",
                        "format": "text",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "userPrototype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.UserPrototype"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            }
        },
        "/validate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "validate auth",
                "parameters": [
                    {
                        "type": "string",
                        "default": "token=xxx",
                        "description": "token",
                        "name": "Cookie",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web_internal_models.UserSwagger"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "web_internal_models.ImageSwagger": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.Item": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "imgurl": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.ItemInOrderSwagger": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "item": {
                    "$ref": "#/definitions/web_internal_models.Item"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "web_internal_models.ItemModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "imgurl": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.ItemPrototype": {
            "type": "object",
            "properties": {
                "imgurl": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.ItemsSwagger": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web_internal_models.Item"
                    }
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "web_internal_models.Order": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/web_internal_models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web_internal_models.OrderStatusSwagger": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.OrderSwagger": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web_internal_models.ItemInOrderSwagger"
                    }
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "web_internal_models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.UserCreds": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.UserPrototype": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                }
            }
        },
        "web_internal_models.UserSwagger": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}